finished right side section
