+++ apps/web/app/components/confirmation-code-input.module.css
@@ -1,0 +1,24 @@
+.inputContainer {
+  display: flex;
+  flex-direction: row;
+}
+
+.inputContainer input {
+    width: 70px;
+    height: 90px !important;
+    font-size: 34px;
+    text-align: center;
+    text-transform: uppercase;
+}
+
+.inputContainer .firstInput {
+    border-top-right-radius: 0 !important;
+    border-bottom-right-radius: 0 !important;
+}
+.inputContainer .lastInput {
+    border-top-left-radius: 0 !important;
+    border-bottom-left-radius: 0 !important;
+}
+.inputContainer .middleInput {
+    border-radius: 0 !important;
+}
