+++ apps/web/app/components/right-section.module.css
@@ -8,7 +8,7 @@
 }
 
 .contentWrapper {
-  width: 400px;
+  width: 520px;
   padding: 0 16px;
   display: flex;
   flex-direction: column;
