+++ apps/web/app/components/submit-email.module.css
@@ -1,0 +1,3 @@
+.button {
+  width: 180px;
+}
